Increase the maximum supported VRFs to 65k
As part of this change, the number of vrfs in vrouter is also made a
module load parameter. Also, during soft reset, memory allocated for
holding the route tables and for vrf statistics are held on instead
of releasing it and allocing it later.

Closes-BUG: #1454497

+++ dp-core/vr_interface.c
@@ -2089,7 +2089,7 @@ vif_vrf_table_get(struct vr_interface *vif, vr_vrf_assign_req *req)
  */
 int
 vif_vrf_table_set(struct vr_interface *vif, unsigned int vlan,
-        short vrf, unsigned short nh_id)
+        int vrf, unsigned int nh_id)
 {
     int ret = 0;
